requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pdfminer.six>=20221105
docx2txt>=0.8
spacy>=3.7.0
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0
numpy>=1.24.0
sqlalchemy>=2.0.0
asyncpg
alembic
apscheduler>=3.10.0
psycopg2-binary
python-dotenv
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0
pydantic-settings>=2.0.0